+++ CHANGELOG.md
@@ -109,6 +109,7 @@
 - Added support for string indexing with `Timedelta` objects.
 - Added support for `Series.dt.total_seconds` method.
 - Added support for `DataFrame.apply(axis=0)`.
+- Added support for `DatetimeIndex.tz_localize` and `DatetimeIndex.tz_convert`.
 
 #### Improvements